Use string templating to build test json
We reuse the same values over and over, so might as well template the
strings. This will make it easier to add or change values in the future
without worrying about pesky copy/paste errors. It also helps in the
special test cases to call out the particular goofy format change from
our default JSON.