Use trait for core processing functionality
We can easily move to a processor per model, as the core trait is where
we can add the logic for looping, error handling and the lowerlevel JSON
bits. The processor class only needs to define the reader (and soon
writer).

I'm not sure on the 'implicit val reader: Reads[A]'. It seems goofy, but
fixed a compiler warning while I figure out what good looks like.